ISSUE #4641 refactor to allow exposure
