+++ backend/src/v5/processors/teamspaces/projects/models/commons/scene.js
@@ -17,9 +17,37 @@
 
 const Scene = {};
 
+const { getMetadataByQuery } = require('../../../../../models/metadata');
 const { getNodesBySharedIds } = require('../../../../../models/scenes');
+const { idTypesToKeys } = require('../../../../../models/metadata.constants');
 
 Scene.getMeshesWithParentIds = (teamspace, project, container, revision, parentIds) => getNodesBySharedIds(teamspace,
 	project, container, revision, parentIds, { _id: 1 });
+
+const getExteralIdNameFromMetadata = (metadata) => {
+	let externalIdName;
+	Object.keys(idTypesToKeys).forEach((name) => {
+		if (idTypesToKeys[name].some((m) => m === metadata[0].metadata[0].key)) {
+			externalIdName = name;
+		}
+	});
+
+	return externalIdName;
+};
+
+Scene.sharedIdsToExternalIds = async (teamspace, container, sharedIds) => {
+	const externalIdKeys = Object.values(idTypesToKeys).flat();
+	const query = { parents: { $in: sharedIds }, 'metadata.key': { $in: externalIdKeys } };
+	const projection = { metadata: { $elemMatch: { $or: externalIdKeys.map((n) => ({ key: n })) } } };
+	const metadata = await getMetadataByQuery(teamspace, container, query, projection);
+
+	if (metadata?.length) {
+		// eslint-disable-next-line no-underscore-dangle
+		const externalIdName = getExteralIdNameFromMetadata(metadata);
+		return { type: externalIdName, values: [...new Set(metadata.map((m) => m.metadata[0].value))] };
+	}
+
+	return undefined;
+};
 
 module.exports = Scene;
